backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
psycopg2-binary>=2.9.7
cx_Oracle==8.3.0
pydantic==2.5.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.8.2
pytz==2023.3
pandas==2.1.3
numpy==1.24.3
requests==2.31.0
aiofiles==23.2.0
wheel
